Integer scaling for resolution scale

+++ src/shared/scaledreso.h
@@ -1,10 +1,12 @@
 #ifndef SCALEDRESO_H
 #define SCALEDRESO_H 1
 
+#include <stdint.h>
+
 #define WIDTH 640
 #define HEIGHT 480
 
-void setscaledreso(float w, float h);
+void setscaledreso(int32_t w, int32_t h);
 void scaledreso(float *w, float *h);
 
 #endif /* SCALEDRESO_H */
